requirements.txt
discord-py-slash-command==3.0.3
discord.py==1.7.3
python-dotenv==0.19.2
pandas==1.3.5
table2ascii==0.1.4
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
requests==2.32.3
numpy==1.26.4
pillow==11.0.0